KAFKA-14523: Move kafka.log.remote classes to storage
